requirements.txt
numpy
Pillow
scikit-learn
matplotlib
seaborn
pandas
scipy
